# Conflicts:
#	pom.xml
#	src/main/java/uz/pdp/hotelsystem/config/SecurityConfig.java
#	src/main/java/uz/pdp/hotelsystem/controller/AuthController.java
#	src/main/java/uz/pdp/hotelsystem/controller/BookingController.java
#	src/main/java/uz/pdp/hotelsystem/controller/EmployeeController.java
#	src/main/java/uz/pdp/hotelsystem/controller/GuestController.java
#	src/main/java/uz/pdp/hotelsystem/controller/HotelController.java
#	src/main/java/uz/pdp/hotelsystem/controller/RoomController.java
#	src/main/java/uz/pdp/hotelsystem/entity/Employee.java
#	src/main/java/uz/pdp/hotelsystem/entity/User.java
#	src/main/java/uz/pdp/hotelsystem/payload/BookingRoomDTO.java
#	src/main/java/uz/pdp/hotelsystem/payload/EmployeeDTO.java
#	src/main/java/uz/pdp/hotelsystem/payload/PaymentDTO.java
#	src/main/resources/application.properties

+++ src/main/java/uz/pdp/hotelsystem/repository/BookingRepository.java
@@ -7,5 +7,5 @@ import uz.pdp.hotelsystem.entity.Payment;
 import java.util.Optional;
 
 
-public interface BookingRepository extends JpaRepository<BookingRoom,Integer> {
+public interface BookingRepository extends JpaRepository<BookingRoom,Long> {
 }
